Used Javascript for Hover Behavior on Profile Icons
I figured out way to use JS to do the hover behavior, however, the implementation isn't too efficient.

Hopefully I can figure out an if statement that'll switch icons depending on the class of the incoming object. Even better if I could use an if statement but instead of switching images, toggle between the colors by accessing the SVG elements.

(I also need to change the hover behavior for the github icon on the navbar from CSS to JS.)

Moving on to spacing the profile icons more appropriately now that I've removed the borders...