Changed Homepage
I managed to make the new homepage mostly responsive. Issues transpire at screen sizes right between small and large screens where the picture doesn't fill its container.

While trying to make that responsive, I realized that with the ipad and ipadpro  sizes, the overall margins shift too much to the left and, with the ipadpro in a vertical orientation, the footer sits directly below the main page content instead of at the bottom of the screen.

Since the ipad sizes are causing problems, I may go into the grid files and edit the sizes for the containers to accomodate..