Fixed: Footer Moving Up Screen Due To Small Amount Of Content
Instead of having a function to check if the two publication sets were both closed, I just added in the empty containers. I quickly realized that once we get enough years under our belt, the issue of not having enough content will be irrelevant, then we'd have an irrelevant function.  No need for that. The next person who maintains this site, when that moment of multiple years comes, they can simply delete the empty container.