Found The Optimal Responsive Layout For Good Spacing
The last item on the website list is almost completed. I messed around with the containers and managed to get a layout for the spacing of the icons and name bit to resuse on the other profile containers. Will need to add extensive comments so I can understand what I did in two weeks and apply the layout to the other containers.